chore: split up gw2data
