+++ wasm_module/src/data.rs
@@ -1,0 +1,5 @@
+pub mod affix;
+pub mod attribute;
+pub mod character;
+pub mod misc;
+pub mod settings;
