+++ wasm_module/src/optimizer_core.rs
@@ -1,11 +1,12 @@
-use wasm_bindgen::{prelude::wasm_bindgen, JsCast, JsValue};
-use web_sys::console;
-
 use crate::{
-    gw2data::{Affix, Attribute, AttributesArray, Character, Combination, Condition},
+    data::{
+        affix::Affix,
+        attribute::Attribute,
+        character::{AttributesArray, Character},
+        settings::{Condition, Settings},
+    },
     result::Result,
-    utils::round_even,
-    utils::{clamp, print_attr},
+    utils::{clamp, round_even},
 };
 use std::cell::RefCell;
 
@@ -49,7 +50,7 @@ pub fn descend_subtree_dfs<F>(
 /// # Arguments
 /// * `chunks` - A vector of vectors of affixes. Each chunk represents a subtree of the affix tree. The chunks are generated by the JS code and distributed to multiple web workers.
 /// * `combinations` - A vector of extras combinations. To calculate the best runes and sigils we must calculate the resulting stats for each combination of extras. Also contains important optimizer settings.
-pub fn start(chunks: &Vec<Vec<Affix>>, combinations: &Vec<Combination>) -> Result {
+pub fn start(chunks: &Vec<Vec<Affix>>, combinations: &Vec<Settings>) -> Result {
     let rankby = combinations[0].rankby;
     // calculate the number of results we need to store;
     let result_num = combinations[0].maxResults; // as f32 / total_threads as f32;
@@ -116,7 +117,7 @@ pub fn start(chunks: &Vec<Vec<Affix>>, combinations: &Vec<Combination>) -> Resul
     return result;
 }
 
-fn test_character(character: &mut Character, settings: &Combination, subtree: &[Affix]) -> bool {
+fn test_character(character: &mut Character, settings: &Settings, subtree: &[Affix]) -> bool {
     // add base attributes from settings to character
     settings.baseAttributes.iter().for_each(|(key, value)| {
         character.base_attributes.set_a(*key, *value);
@@ -141,7 +142,7 @@ fn test_character(character: &mut Character, settings: &Combination, subtree: &[
     return update_attributes(character, settings, false);
 }
 
-fn update_attributes(character: &mut Character, settings: &Combination, no_rounding: bool) -> bool {
+fn update_attributes(character: &mut Character, settings: &Settings, no_rounding: bool) -> bool {
     calc_stats(character, settings, no_rounding);
     //print_attr(&character.attributes);
 
@@ -163,7 +164,7 @@ fn update_attributes(character: &mut Character, settings: &Combination, no_round
     return true;
 }
 
-fn calc_stats(character: &mut Character, settings: &Combination, no_rounding: bool) {
+fn calc_stats(character: &mut Character, settings: &Settings, no_rounding: bool) {
     // move base attributes to attributes as default
     // not sure which method is faster, but I think the for loop is faster:
     // 1. for loop
@@ -284,7 +285,7 @@ fn calc_stats(character: &mut Character, settings: &Combination, no_rounding: bo
     }
 }
 
-pub fn calc_power(character: &mut Character, settings: &Combination) -> f32 {
+pub fn calc_power(character: &mut Character, settings: &Settings) -> f32 {
     let attributes = &mut character.attributes;
     let mods = &settings.modifiers;
 
@@ -346,7 +347,7 @@ fn condition_damage_tick(
 
 pub fn calc_condi(
     character: &mut Character,
-    settings: &Combination,
+    settings: &Settings,
     relevant_conditions: &Vec<Condition>,
 ) -> f32 {
     let attributes = &mut character.attributes;
